requirements.txt
timm==0.9.5
pytorch_msssim==1.0.0

omegaconf==2.3.0
torchmetrics==0.10.3

tensorboard==2.11.0
wandb==0.15.8
matplotlib==3.5.3
plotly==5.13.1
viser==0.2.1

imageio
imageio-ffmpeg
scikit-image==0.20.0
opencv-python

open3d==0.16.0
pyquaternion==0.9.9
chumpy
numpy==1.23.1
kornia==0.7.2

tqdm
gdown

nerfview==0.0.3
lpips==0.1.4
tensorly